BUG: Drive letter is ignored on save file
FileDialog.file_selected does not contain the selected drive letter. If you select another drive on save it will obviously save relative to the application dir. But the file_selected signal holds the complete path as return value.
Added a local variable to store the return value of the signal and use it as parameter for save_file().